% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpic.R
\name{rpic_knitr_engine}
\alias{rpic_knitr_engine}
\title{knitr engine for \code{{rpic}} chunks}
\usage{
rpic_knitr_engine(options)
}
\arguments{
\item{options}{knitr chunk options.}
}
\description{
Register with \code{\link[=rpic_register_knitr]{rpic_register_knitr()}}, then write pic code in a chunk:

\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{rpic, circuits=TRUE\}
A:(0,0); B:(2,0)
resistor(A,B)
```
}\if{html}{\out{</div>}}
}
