% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpic.R
\name{rpic_manifest}
\alias{rpic_manifest}
\title{Compile to a JSON \verb{\{svg, animations\}} bundle (as a string)}
\usage{
rpic_manifest(src, circuits = FALSE)
}
\arguments{
\item{src}{pic source code.}

\item{circuits}{load the native circuit-element library.}
}
\value{
a JSON string.
}
\description{
Compile to a JSON \verb{\{svg, animations\}} bundle (as a string)
}
\examples{
rpic_manifest('box; animate last box with "pop"')
}
