% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpic.R
\name{rpic_svg}
\alias{rpic_svg}
\title{Render pic source to an SVG string}
\usage{
rpic_svg(src, circuits = FALSE)
}
\arguments{
\item{src}{pic source code.}

\item{circuits}{load the native circuit-element library.}
}
\value{
an SVG string.
}
\description{
Render pic source to an SVG string
}
